GIF XMP read and write support (#119)
### Overview
This pull request introduces the ability to read and write XMP metadata
from GIF files by implementing an `ImageParser` for the [GIF
spec](https://www.w3.org/Graphics/GIF/spec-gif89a.txt) and encoding XMP
metadata as an Application Extension Block as described in the [XMP file
embedding
spec](https://github.com/adobe/XMP-Toolkit-SDK/blob/main/docs/XMPSpecificationPart3.pdf).

### Changes
The following classes have been added, mostly following the pattern of
the classes in the PNG package:
- `GifImageParser`, an implementation of `ImageParser` for GIF files
- `GifMetadata`, an implementation of `MetadataExtractor` for GIF files
- `GifUpdater`, an implementation of `MetadataUpdater` for GIF files
- `GifWriter` for writing `GifChunks` back into binary
- A bunch of helper classes and enums to describe the GIF format as a
series of chunks

Tests have been written for the new functionality, again mostly
following the pattern for PNG files.

Additionally, I've included
- Some minor refactoring of `ByteWriter` classes and extensions
- Additional parameters to `AbstractUpdaterTest` to skip test cases for
metadata updates that don't apply to the GIF format: image rotation and
thumbnails

Happy to receive feedback on these changes, I'm keen to see this library
fleshed out more to support a wider range of XMP embedded formats.

---------

+++ src/commonMain/kotlin/com/ashampoo/kim/output/BufferByteWriter.kt
@@ -23,12 +23,12 @@ internal class BufferByteWriter(
     private var index: Int
 ) : ByteWriter {
 
-    override fun write(byte: Int) {
+    override fun write(byte: Byte) {
 
         if (index >= buffer.size)
             throw ImageWriteException("Buffer overflow.")
 
-        buffer[index++] = byte.toByte()
+        buffer[index++] = byte
     }
 
     override fun write(byteArray: ByteArray) {
